+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Unreleased
 ==========
 - lib.hardware.tispeech: Init
 - lib.hardware.tispeech: Add TMS0280 constants
+- lib.hardware.tispeech: Add SpeakAndSpellROM
 
 0.9.0 - 2021-01-14
 ==================